On branch main  Your branch is ahead of 'origin/main' by 3 commits.    (use "git push" to publish your local commits)
 Changes to be committed:
	modified:   src/empresa_agro/Almacen.java
	modified:   src/empresa_agro/Main.java
	modified:   src/empresa_agro/Producto.java

+++ src/empresa_agro/Almacen.java
@@ -16,7 +16,7 @@ public class Almacen {
         } else {
 
             int numKilos;
-            int precioKilos;
+            double precioKilos;
 
             producto.setAlmacenado(true);
